+++ examples/list-servers.rs
@@ -26,9 +26,8 @@ fn main() {
     let identity = Identity::from_env()
         .expect("Failed to create an identity provider from the environment");
     let session = Session::new(identity);
-
-    let manager = compute::servers::manager(&session);
-    let servers = manager.list().expect("Cannot list servers");
+    let servers = compute::servers::manager(&session).list()
+        .expect("Cannot list servers");
     for s in &servers {
         println!("ID = {}, Name = {}", s.id(), s.name());
     }