+++ src/utils.rs
@@ -23,7 +23,6 @@ use hyper::Client;
 use hyper::net::HttpsConnector;
 #[cfg(feature = "tls")]
 use hyper_rustls::TlsClient;
-use uuid::Uuid;
 
 use super::ApiResult;
 
@@ -103,36 +102,6 @@ impl<K: Hash + Eq, V: Clone> MapCache<K, V> {
 }
 
 
-/// Something that can be converted to an ID.
-pub trait IntoId {
-    /// Convert a value into an ID.
-    fn into_id(self) -> String;
-}
-
-impl IntoId for Uuid {
-    fn into_id(self) -> String {
-        self.to_string()
-    }
-}
-
-impl IntoId for String {
-    fn into_id(self) -> String {
-        self
-    }
-}
-
-impl<'a> IntoId for &'a String {
-    fn into_id(self) -> String {
-        self.clone()
-    }
-}
-
-impl<'a> IntoId for &'a str {
-    fn into_id(self) -> String {
-        String::from(self)
-    }
-}
-
 pub mod url {
     //! Handy primitives for working with URLs.
 
@@ -155,7 +124,8 @@ pub mod url {
 
     #[inline]
     #[allow(unused_results)]
-    pub fn extend(mut url: Url, segments: &[&str]) -> Url {
+    pub fn extend<I>(mut url: Url, segments: I) -> Url
+            where I: IntoIterator, I::Item: AsRef<str> {
         url.path_segments_mut().unwrap().pop_if_empty().extend(segments);
         url
     }