+++ CHANGES.md
@@ -4,6 +4,9 @@ ChangeLog
 next
 ----
 
+* Dropped ServiceApi struct in favor of using Session directly.
+* Rename compute::{ComputeV2Type -> V2Type};
+
 0.0.8
 -----